// packages/shared/src/types/__generated/kysely.type.ts
export interface MetadataDatabase {
  connection_id: string;
  description: string | null;
  display_name: string | null;
  link: string | null;
  organization_id: string;
  raw_name: string;
  slug: string;
  summary: string | null;
}